chore(service): restructure project files for better maintainability
