+++ service/main.py
@@ -3,9 +3,9 @@ from datetime import datetime, timezone
 from fastapi import FastAPI, HTTPException, Response, status
 from prometheus_fastapi_instrumentator import Instrumentator
 
-from .metadata import Metadata
-from .models.model import RMRModel
-from .schemas import HealthStatus, InputData, OutputData
+from metadata import Metadata
+from models.model import RMRModel
+from schemas import HealthStatus, InputData, OutputData
 
 rmr_model = RMRModel()
 start_time_utc = datetime.now(timezone.utc)  # Store the time the api starts
