+++ service/models/equations.py
