+++ service/tests/conftest.py
@@ -1,9 +1,17 @@
 import numpy as np
 import pytest
 
-from ..models.equations import Equations
+from ..main import app
+from fastapi.testclient import TestClient
+from ..models.model import Equations
 from ..models.model import RMRModel
 from ..models.time_projection import TimeProjection
+
+
+@pytest.fixture
+def client() -> TestClient:
+    client = TestClient(app)
+    return client
 
 
 @pytest.fixture
