+++ service/tests/test_equations.py
